fix checks

+++ predicators/planning.py
@@ -1046,7 +1046,7 @@ def _update_sas_file_with_costs(
 def fd_plan_from_sas_file(
     sas_file: str, timeout_cmd: str, timeout: float, exec_str: str,
     alias_flag: str, start_time: float, objects: List[Object],
-    init_atoms: Set[GroundAtom], nsrts: Set[NSRT], max_horizon: int
+    init_atoms: Set[GroundAtom], nsrts: Set[NSRT], max_horizon: float
 ) -> Tuple[List[_GroundNSRT], List[Set[GroundAtom]],
            Metrics]:  # pragma: no cover
     """Given a SAS file, runs search on it to generate a plan."""
@@ -1137,7 +1137,7 @@ def _sesame_plan_with_fast_downward(
     while True:
         skeleton, atoms_sequence, metrics = fd_plan_from_sas_file(
             sas_file, timeout_cmd, timeout, exec_str, alias_flag, start_time,
-            objects, init_atoms, nsrts, max_horizon)
+            objects, init_atoms, nsrts, float(max_horizon))
         # Run low-level search on this skeleton.
         low_level_timeout = timeout - (time.perf_counter() - start_time)
         try:
@@ -1247,7 +1247,7 @@ def run_task_plan_once(
 
         plan, atoms_seq, metrics = fd_plan_from_sas_file(
             sas_file, timeout_cmd, timeout, exec_str, alias_flag, start_time,
-            list(objects), init_atoms, nsrts, max_horizon)
+            list(objects), init_atoms, nsrts, float(max_horizon))
     else:
         raise ValueError("Unrecognized sesame_task_planner: "
                          f"{CFG.sesame_task_planner}")
